+++ 3rd/math/src/fabs.c
@@ -1,8 +1,9 @@
-#include <math.h>
+#include "math.h"
+#include <stdint.h>
 
 double fabs(double x)
 {
-	union {double f; unsigned long long i;} u = {x};
+	union {double f; uint64_t i;} u = {x};
 	u.i &= ULLONG_NSHIFT/2;
 	return u.f;
 }
