+++ 3rd/math/src/fmod.c
@@ -1,16 +1,17 @@
-#include <math.h>
+#include "math.h"
+#include <stdint.h>
 
 double fmod(double x, double y)
 {
-	union {double f; unsigned long long i;} ux = {x}, uy = {y};
+	union {double f; uint64_t i;} ux = {x}, uy = {y};
 	int ex = ux.i>>52 & 0x7ff;
 	int ey = uy.i>>52 & 0x7ff;
 	int sx = ux.i>>63;
-	unsigned long long i;
+	uint64_t i;
 
 	/* in the followings uxi should be ux.i, but then gcc wrongly adds */
 	/* float load/store to inner loops ruining performance and code size */
-	unsigned long long uxi = ux.i;
+	uint64_t uxi = ux.i;
 
 	if (uy.i<<1 == 0 || isnan(y) || ex == 0x7ff)
 		return (x*y)/(x*y);
@@ -57,11 +58,11 @@ double fmod(double x, double y)
 	/* scale result */
 	if (ex > 0) {
 		uxi -= 1ULL << 52;
-		uxi |= (unsigned long long)ex << 52;
+		uxi |= (uint64_t)ex << 52;
 	} else {
 		uxi >>= -ex + 1;
 	}
-	uxi |= (unsigned long long)sx << 63;
+	uxi |= (uint64_t)sx << 63;
 	ux.i = uxi;
 	return ux.f;
 }
