+++ 3rd/math/src/modf.c
@@ -1,9 +1,9 @@
-#include "math.h"
+#include "libm.h"
 
 double modf(double x, double *iptr)
 {
-	union {double f; unsigned long long i;} u = {x};
-	unsigned long long mask;
+	union {double f; uint64_t i;} u = {x};
+	uint64_t mask;
 	int e = (int)(u.i>>52 & 0x7ff) - 0x3ff;
 
 	/* no fractional part */
